dataset is not an instance of OrderedDict anymore if has attributes - data (ordered Dict of dimarray.dataset.Variable instances) - axes (common axes)
It has methods __getitem__, __setitem__ and so on to mimic a dictionary.

The main difference with a dict is the copy/ref behaviour, which is
not systematic. For instance, setting metadata d['var'].units = 'm'
since d['var'] will create a new DimArray instance on the fly.
will fail. Can still do d['var'].data.units = 'm'.

The advantage of this approach is that there is truely only one `axes`
instance. Modifying it will affect all dimarrays simultaneously.