changed datasets structure : now all element in a Dataset truely share their axes. When a new DimArray is inserted in a Dataset, its axes are replaced with any pre-exising, like-named axis from the Dataset. That way, any operation on a Dataset axis is felt for all elements of the Dataset which share that axis.
Note that a shallow copy (axes, metadata) is performed *prior* to inserting
the array in the data set.

Canonical __getitem__ method is used, without any further copy
when accessing a Dataset element. As a result, it is possible
to perform changes on single elements attributes from within
the dataset (e.g. metadata), which is more pythonic as far as
a dictionary is concerned.

An important consequence of this reference (instead of copy) access
is that after having been inserted in a Dataset, any operation on
a DimArray axes will further affect all other DimArray of the dataset
which share that axis.

If is of course possible to perform a copy of the axes (or of the whole
dimarray) if this behaviour is not desired.